--- src/ui/common/src/handlers/parameters/DagId.ts
@@ -1,3 +1,0 @@
-export type DagIdParameter = {
-  dagId: string;
-};
