--- src/ui/common/src/handlers/parameters/DagResultId.ts
@@ -1,3 +1,0 @@
-export type DagResultIdParameter = {
-  dagResultId: string;
-};
