--- src/ui/common/src/handlers/parameters/NodeId.ts
@@ -1,3 +1,0 @@
-export type NodeIdParameter = {
-  nodeId: string;
-};
