--- src/ui/common/src/handlers/parameters/NodeResultId.ts
@@ -1,3 +1,0 @@
-export type NodeResultIdParameter = {
-  nodeResultId: string;
-};
