+++ src/ui/common/src/handlers/parameters/Path.ts
@@ -1,0 +1,19 @@
+export type DagIdParameter = {
+  dagId: string;
+};
+
+export type DagResultIdParameter = {
+  dagResultId: string;
+};
+
+export type NodeIdParameter = {
+  nodeId: string;
+};
+
+export type NodeResultIdParameter = {
+  nodeResultId: string;
+};
+
+export type WorkflowIdParameter = {
+  workflowId: string;
+};
