--- src/ui/common/src/handlers/parameters/WorkflowId.ts
@@ -1,3 +1,0 @@
-export type WorkflowIdParameter = {
-  workflowId: string;
-};
