+++ src/ui/common/src/handlers/v2/DagResultGet.ts
@@ -1,9 +1,8 @@
 // This file should map exactly to
 // src/golang/cmd/server/handler/v2/dag_result_get.go
 
-import { APIKeyParameter } from '../parameters/ApiKey';
-import { DagResultIdParameter } from '../parameters/DagResultId';
-import { WorkflowIdParameter } from '../parameters/WorkflowId';
+import { APIKeyParameter } from '../parameters/Header';
+import { DagResultIdParameter, WorkflowIdParameter } from '../parameters/Path';
 import { DagResultResponse } from '../responses/Workflow';
 
 export type DagResultGetRequest = APIKeyParameter &
