+++ src/ui/common/src/handlers/v2/NodeArtifactGet.ts
@@ -1,10 +1,12 @@
 // This file should map exactly to
 // src/golang/cmd/server/handler/v2/node_artifact_get.go
 
-import { APIKeyParameter } from '../parameters/ApiKey';
-import { DagIdParameter } from '../parameters/DagId';
-import { NodeIdParameter } from '../parameters/NodeId';
-import { WorkflowIdParameter } from '../parameters/WorkflowId';
+import { APIKeyParameter } from '../parameters/Header';
+import {
+  DagIdParameter,
+  NodeIdParameter,
+  WorkflowIdParameter,
+} from '../parameters/Path';
 import { ArtifactResponse } from '../responses/Node';
 
 export type NodeArtifactGetRequest = APIKeyParameter &
