+++ src/ui/common/src/handlers/v2/NodesResultsGet.ts
@@ -1,9 +1,8 @@
 // This file should map exactly to
 // src/golang/cmd/server/handler/v2/nodes_results_get.go
 
-import { APIKeyParameter } from '../parameters/ApiKey';
-import { DagResultIdParameter } from '../parameters/DagResultId';
-import { WorkflowIdParameter } from '../parameters/WorkflowId';
+import { APIKeyParameter } from '../parameters/Header';
+import { DagResultIdParameter, WorkflowIdParameter } from '../parameters/Path';
 import { NodeResultsResponse } from '../responses/Node';
 
 export type NodesResultsGetRequest = APIKeyParameter &
