+++ src/ui/common/src/handlers/v2/WorkflowGet.ts
@@ -1,7 +1,7 @@
 // This file should map exactly to
 // src/golang/cmd/server/handler/v2/workflow_get.go
-import { APIKeyParameter } from '../parameters/ApiKey';
-import { WorkflowIdParameter } from '../parameters/WorkflowId';
+import { APIKeyParameter } from '../parameters/Header';
+import { WorkflowIdParameter } from '../parameters/Path';
 import { WorkflowResponse } from '../responses/Workflow';
 
 export type WorkflowGetRequest = APIKeyParameter & WorkflowIdParameter;
